requirements.txt
requests
pandas
sqlalchemy
pymysql
python-dotenv
xmltodict